- Finalizado implementação das controllers

+++ Desafio/src/Desafio.Application/Services/PersonService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Desafio.Domain;
+using Microsoft.AspNetCore.Identity;
 using System.Net.Http.Headers;
 
 namespace Desafio.Application;
@@ -28,6 +29,18 @@ public class PersonService : ServiceBase, IPersonService
 
         return result;
     }
+    public async Task<IEnumerable<PersonResponse>> GetAllClientAsync()
+    {
+        var result = _mapper.Map<IEnumerable<PersonResponse>>(await _personRepository.GetAllClientAsync());
+
+        if (result == null)
+        {
+            Notificate("No client was found");
+            return null;
+        }
+
+        return result;
+    }
 
     public async Task<PersonResponse> GetByIdAsync(Guid id)
     {
@@ -36,6 +49,19 @@ public class PersonService : ServiceBase, IPersonService
         if (person == null)
         {
             Notificate("No person was found.");
+            return null;
+        }
+
+        return _mapper.Map<PersonResponse>(person);
+    }
+
+    public async Task<PersonResponse> GetClientByIdAsync(Guid id)
+    {
+        var person = await _personRepository.GetClientByIdAsync(id);
+
+        if (person == null)
+        {
+            Notificate("No client was found.");
             return null;
         }
 
@@ -57,7 +83,10 @@ public class PersonService : ServiceBase, IPersonService
 
     public async Task<PersonResponse> InsertAsync(InsertPersonRequest personRequest)
     {
+        
         var person = _mapper.Map<Person>(personRequest);
+
+        person.Document = OnlyDocumentNumbers(personRequest.Document);
 
         if (!await ExecuteValidationAsync(new PersonValidator(this), person))
         {
@@ -80,6 +109,13 @@ public class PersonService : ServiceBase, IPersonService
             Notificate("No person was found.");
             return null;
         }
+
+        if (!await ExecuteValidationAsync(new RemovePersonValidator(this), person))
+        {
+            return null;
+        }
+
+
         await _personRepository.RemoveAsync(id);
 
         return null;
@@ -90,19 +126,17 @@ public class PersonService : ServiceBase, IPersonService
 
         var existingperson = await _personRepository.GetByIdAsync(personRequest.Id);
 
-        if (existingperson != null)
+        if (existingperson == null)
         {
             Notificate("No person was found.");
             return null;
         }
 
-        existingperson.Name = personRequest.Name;
-        existingperson.Document = personRequest.Document;
-        existingperson.City = personRequest.City;
-        existingperson.Enable = personRequest.Enable;
-        existingperson.CanBuy = personRequest.CanBuy;
-        existingperson.Notes = personRequest.Notes;
-        existingperson.AlternativeCode = personRequest.AlternativeCode;
+        existingperson.Name = personRequest.Name ?? existingperson.Name;
+        existingperson.Document = personRequest.Document ?? existingperson.Document;
+        existingperson.City = personRequest.City ?? existingperson.City;
+        existingperson.Notes = personRequest.Notes ?? existingperson.Notes;
+        existingperson.AlternativeCode = personRequest.AlternativeCode ?? existingperson.AlternativeCode;
 
 
         if (!await ExecuteValidationAsync(new PersonValidator(this), existingperson))
@@ -154,6 +188,38 @@ public class PersonService : ServiceBase, IPersonService
 
         return personResponse;
     }
+    #endregion
 
+    #region Validation Methods
+    public async Task<bool> DocumentAlreadyExistsAsync(string document)
+    {
+        return await _personRepository.DocumentAlreadyExistsAsync(document);
+    }
+    public async Task<bool> AlternativeCodeAlreadyExistsAsync(string alternativeCode)
+    {
+        return await _personRepository.AlternativeCodeAlreadyExistsAsync(alternativeCode);
+    }
+    public async Task<bool> PersonCanBuyAsync(Guid id)
+    {
+        return await _personRepository.PersonCanBuyAsync(id);
+    }
+
+    public bool IsValidDocument(string document, bool canBeNullOrEmpty = false)
+    {
+        if(canBeNullOrEmpty && string.IsNullOrWhiteSpace(document))
+        {
+            return true;
+        }
+        string documentNumber = OnlyDocumentNumbers(document);
+        bool validLength = documentNumber.Length == 11 || documentNumber.Length == 14;
+
+        if (string.IsNullOrWhiteSpace(documentNumber) || HasRepeatedValues(documentNumber) || !validLength)
+        {
+            return false;
+        }
+
+        return true;
+
+    }
     #endregion
 }
